feat(MFA): Added support for custom verification checks
Added support for custom MFA verification checks, such as receiving MFA
codes using e-mail and SMS. These usually are valid for a longer period,
so the default TOTP validator cannot be used.

The current version of rest-secure did not allow such checks, since if
it detected a `verificationCode` and the user's `isMfaConfigured()`
returned true, it would throw a bad credentials exception.

Since these checks are usually in some part dependent on the implementing
application, I've chosen not to implement the checks in the library,
but made it easy to extend the default flow by adding support for custom
checks.