FormUtil: Do not throw for invalid request body
When a HTTP request with an invalid request body (XML, HTML) is sent
through either RestAuthenticationFilter or RestAccessDeniedHandler,
FormUtil would fail with an IllegalStateException.

This cluttered our logging / monitoring, because it often happens due
to automated attempts to get access to systems by bots / hackers.

Instead of throwing an exception, we now log a warning and return an
empty form, so the mis-use of systems is still possible to detect,
but no longer triggers error alerting rules.